improved code but it is not good yet
improved the very old code a bit, although the A3C algorithms is not yet well-implemented. Importantly, there is no exploration scheme implemented yet, such as an epsilon-greedy approach or entropy regularization. Due to this, the performance is still quite poor. All of this should be easy to implement, but I'll leave it here!